adress feedback (I believe that this captures all changes requested)
