+++ src/duplex.rs
@@ -1,17 +1,6 @@
-//! Duplex audio stream support with synchronized input/output.
-//!
-//! This module provides types for building duplex (simultaneous input/output) audio streams
-//! with hardware clock synchronization.
-//!
-//! See `examples/duplex_feedback.rs` for a working example.
-
 use crate::{ChannelCount, InputStreamTimestamp, OutputStreamTimestamp, SampleRate};
 
-/// Information passed to duplex callbacks.
-///
-/// This contains timing information for the current audio buffer, combining
-/// both input and output timing. A duplex stream has a single callback invocation
-/// that provides synchronized input and output data.
+// Timing information for a duplex callback, combining input and output timestamps.
 #[derive(Clone, Copy, Debug, PartialEq, Eq, Hash)]
 pub struct DuplexCallbackInfo {
     input_timestamp: InputStreamTimestamp,
@@ -19,10 +8,6 @@ pub struct DuplexCallbackInfo {
 }
 
 impl DuplexCallbackInfo {
-    /// Create a new DuplexCallbackInfo.
-    ///
-    /// Note: Both timestamps will share the same `callback` instant since there is
-    /// only one callback invocation for a duplex stream.
     pub fn new(
         input_timestamp: InputStreamTimestamp,
         output_timestamp: OutputStreamTimestamp,
@@ -33,36 +18,19 @@ impl DuplexCallbackInfo {
         }
     }
 
-    /// The timestamp for the input portion of the duplex stream.
-    ///
-    /// Contains the callback instant and when the input data was captured.
     pub fn input_timestamp(&self) -> InputStreamTimestamp {
         self.input_timestamp
     }
 
-    /// The timestamp for the output portion of the duplex stream.
-    ///
-    /// Contains the callback instant and when the output data will be played.
     pub fn output_timestamp(&self) -> OutputStreamTimestamp {
         self.output_timestamp
     }
 }
 
-/// Configuration for a duplex audio stream.
-///
-/// Unlike separate input/output streams, duplex streams require matching
-/// configuration for both directions since they share a single device context.
 #[derive(Clone, Debug, Eq, PartialEq)]
 pub struct DuplexStreamConfig {
-    /// Number of input channels.
     pub input_channels: ChannelCount,
-
-    /// Number of output channels.
     pub output_channels: ChannelCount,
-
-    /// Sample rate in Hz.
     pub sample_rate: SampleRate,
-
-    /// Requested buffer size in frames.
     pub buffer_size: crate::BufferSize,
 }
